Move engine related structs to engine module
Change-type: patch
